Didn't convert

+++ preprocess.py
@@ -61,10 +61,12 @@ class PreprocessDataset(Dataset):
             speech_style = contents["화자정보"]["SpeechStyle"]
             character = contents["화자정보"]["Character"]
             character_emotion = contents["화자정보"]["CharacterEmotion"]
-            begin_time = (
+            begin_time = int(
                 float(contents["기타정보"]["SpeechStart"]) * self.cfg.data.sample_rate
             )
-            end_time = contents["기타정보"]["SpeechEnd"] * self.cfg.data.sample_rate
+            end_time = int(
+                float(contents["기타정보"]["SpeechEnd"]) * self.cfg.data.sample_rate
+            )
         if not wav_path.exists():
             print(f"Audio file {wav_path} does not exist")
             return None
